refactor: remove compat shims; update all callers to import from actual modules
sprint/__init__.py: trim to 4-symbol public API (run_sprint, SprintManifest,
SprintResult, load_sprint_manifest). All previous re-exports removed.
- cli/run.py: import _triage_spec from sprint.dag directly
- tests/test_sprint.py: import StoryDAG/StoryTriage/build_dag/_triage_spec from
  sprint.dag, _build_task_from_story from sprint.manifest,
  _classify_and_record from sprint.runner
- tests/test_daemon.py: import/patch run_task via sprint.runner

coordinator/notify.py: remove re-export blocks for ntfy_client, remote_gates,
pending_hitl symbols. notify.py now contains only the functions that belong there.

coordinator/engine.py: import _ntfy_poll_reply/_ntfy_publish/_ntfy_reply_url
from ntfy_client, _pending_plan_review from pending_hitl, and
_plan_review_remote/_remote_human_review from remote_gates directly.

coordinator/phases.py: import _escalate_gate_remote/_remote_human_review from
remote_gates; _pending_escalate_gate/_pending_human_review from pending_hitl.

tests/test_coord_notify.py: import _ntfy_poll_reply/_ntfy_reply_url/_ntfy_publish
from ntfy_client; _plan_review_remote from remote_gates directly.
tests/test_coordinator.py: import _begin_run_log_tee/_end_run_log_tee from
coordinator.log_tee (not engine).

1413 tests passing, make gate PASS.
