+++ src/theforge/coordinator/notify.py
@@ -15,24 +15,7 @@ from theforge.review import ReviewResult
 from theforge.task import TaskStory as TaskSpec  # noqa: F401
 
 from . import util as _cu
-from .ntfy_client import (  # noqa: F401
-    _BLOCKING_POLL_CHUNK,
-    _ntfy_poll_escalate_reply,
-    _ntfy_poll_plan_reply,
-    _ntfy_poll_reply,
-    _ntfy_publish,
-    _ntfy_reply_url,
-)
-from .pending_hitl import (  # noqa: F401
-    _pending_escalate_gate,
-    _pending_human_review,
-    _pending_plan_review,
-)
-from .remote_gates import (  # noqa: F401
-    _escalate_gate_remote,
-    _plan_review_remote,
-    _remote_human_review,
-)
+from .ntfy_client import _ntfy_publish
 
 if TYPE_CHECKING:
     from . import state as _cs
