+++ src/theforge/sprint/__init__.py
@@ -1,29 +1,11 @@
 """theforge.sprint — sprint mode package."""
 
-from .dag import StoryDAG, StoryTriage, _run_gate, _triage_spec, build_dag  # noqa: F401
-from .manifest import (  # noqa: F401
-    SprintManifest,
-    SprintResult,
-    _build_task_from_story,
-    load_sprint_manifest,
-)
-from .runner import (  # noqa: F401
-    _classify_and_record,
-    _merge_branch,
-    _notify,
-    _ntfy_publish,
-    run_from_dev,
-    run_from_review,
-    run_sprint,
-    run_task,
-)
+from .manifest import SprintManifest, SprintResult, load_sprint_manifest
+from .runner import run_sprint
 
 __all__ = [
     "SprintManifest",
     "SprintResult",
-    "StoryDAG",
-    "StoryTriage",
-    "build_dag",
     "load_sprint_manifest",
     "run_sprint",
 ]
